// src/app/core/models/cards.dashboard.model.ts
export class RequisicaoCardsModel {
  id_empresa: number;
}

export class RespostaCardsModel {
  success: string;
  msg: string;
  id_empresa: string;
  qtde_email: number;
  total_email: number;
  qtde_sms: number;
  total_sms: number;
  qtde_whatsapp: number;
  total_whatsapp: number;
  total_utilizado: number;
  saldo: number;
}